wtp/webrtc/RTCPeerConnection-setRemoteDescription.html

<!doctype html>
<meta charset="utf-8">
<title>RTCPeerConnection.prototype.setRemoteDescription</title>
<script src="/resources/webrtc-ie-shim.js"></script><script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script>
  'use strict';
  
  // Test is based on the following editor draft:
  // https://w3c.github.io/webrtc-pc/archives/20170605/webrtc.html
  
  // The following helper functions are called from RTCPeerConnection-helper.js:
  //   generateOffer()
  //   generateAnswer()
  //   assert_session_desc_not_equals()
  //   assert_session_desc_equals()
  //   test_state_change_event()
  
  /*
    4.3.2.  Interface Definition
      [Constructor(optional RTCConfiguration configuration)]
      interface RTCPeerConnection : EventTarget {
        Promise<void>                      setRemoteDescription(
            RTCSessionDescriptionInit description);
         readonly attribute RTCSessionDescription? remoteDescription;
        readonly attribute RTCSessionDescription? currentRemoteDescription;
        readonly attribute RTCSessionDescription? pendingRemoteDescription;
        ...
      };
     4.6.2.  RTCSessionDescription Class
      dictionary RTCSessionDescriptionInit {
        required RTCSdpType type;
                 DOMString  sdp = "";
      };
     4.6.1.  RTCSdpType
      enum RTCSdpType {
        "offer",
        "pranswer",
        "answer",
        "rollback"
      };
   */
  
  /*
    4.6.1.  enum RTCSdpType
   */
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
  
    // SDP is validated after WebIDL validation
    return promise_rejects(t, new TypeError(), pc.setRemoteDescription({
      type: 'bogus',
      sdp: 'bogus'
    }));
  }, 'setRemoteDescription with invalid type and invalid SDP should reject with TypeError');
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
  
    // SDP is validated after validating type
    return promise_rejects(t, 'InvalidStateError', pc.setRemoteDescription({
      type: 'answer',
      sdp: 'invalid'
    }));
  }, 'setRemoteDescription() with invalid SDP and stable state should reject with InvalidStateError');
  
  /* Operations after returning to stable state */
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    var pc2 = new RTCPeerConnection();
  
    test_state_change_event(t, pc, ['have-remote-offer', 'stable', 'have-remote-offer']);
  
    return pc2.createOffer({ offerToReceiveAudio: true }).then(function (offer1) {
      return pc.setRemoteDescription(offer1).then(function () {
        return pc.createAnswer();
      }).then(function (answer) {
        return pc.setLocalDescription(answer);
      }).then(function () {
        return pc2.createOffer({ offerToReceiveVideo: true });
      }).then(function (offer2) {
        return pc.setRemoteDescription(offer2).then(function () {
          assert_equals(pc.signalingState, 'have-remote-offer');
          assert_session_desc_not_equals(offer1, offer2);
          assert_session_desc_equals(pc.remoteDescription, offer2);
          assert_session_desc_equals(pc.currentRemoteDescription, offer1);
          assert_session_desc_equals(pc.pendingRemoteDescription, offer2);
        });
      });
    });
  }, 'Calling setRemoteDescription() again after one round of remote-offer/local-answer should succeed');
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
  
    test_state_change_event(t, pc, ['have-local-offer', 'stable', 'have-remote-offer']);
  
    return pc.createOffer({ offerToReceiveAudio: true }).then(function (offer) {
      return pc.setLocalDescription(offer).then(function () {
        return generateAnswer(offer);
      });
    }).then(function (answer) {
      return pc.setRemoteDescription(answer).then(function () {
        return generateOffer({ pc: pc, data: true });
      }).then(function (offer) {
        return pc.setRemoteDescription(offer).then(function () {
          assert_equals(pc.signalingState, 'have-remote-offer');
          assert_session_desc_equals(pc.remoteDescription, offer);
          assert_session_desc_equals(pc.currentRemoteDescription, answer);
          assert_session_desc_equals(pc.pendingRemoteDescription, offer);
        });
      });
    });
  }, 'Switching role from offerer to answerer after going back to stable state should succeed');
  
  /*
    TODO
    4.3.2.  setRemoteDescription
      - If an a=identity attribute is present in the session description, the browser
        validates the identity assertion.
   */

</script>
